+++ src/Mutable/Generic/ListCollection.php
@@ -16,6 +16,7 @@ class ListCollection extends \MF\Collection\Mutable\ListCollection implements IL
         TypeValidator::TYPE_FLOAT,
         TypeValidator::TYPE_BOOL,
         TypeValidator::TYPE_ARRAY,
+        TypeValidator::TYPE_CALLABLE,
         TypeValidator::TYPE_OBJECT,
         TypeValidator::TYPE_INSTANCE_OF,
     ];