+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Allow `callable` type in `Generic` `Lists` and `Maps`
 
 ## 3.14.0 - 2018-10-10
 - Expand `toStringForUrl` method not to quote string containing `_`, `-`, `.` and ` `